[refactor] Split web app python files into their own folder
