+++ backend/src/webApp/webApp.py
@@ -24,7 +24,7 @@ from flask_socketio import SocketIO
 from flask_login import login_user, current_user, login_required, logout_user
 
 #--------------------------------OUR DEPENDENCIES--------------------------------#
-from . import utils
+from .. import utils
 from .webAppUsers import User, UserManager
 from .webAppRegistration import RegistrationForm
 from .webAppLoginForm import LoginForm
